no x spacing between textfields now
